Move db to its own lib
