+++ app/actix/Cargo.toml
@@ -4,6 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+type_ahead_db = { path = "../db" }
+
 actix-web = "4"
 r2d2 = "0.8"
 r2d2_sqlite = "0.21"
