+++ app/actix/src/main.rs
@@ -2,9 +2,7 @@ use actix_web::{error, web, App, HttpResponse, HttpServer, Responder};
 use r2d2_sqlite::{self, SqliteConnectionManager};
 use serde::Deserialize;
 
-// TODO: crate setup
-mod db;
-use db::{Pool, Query};
+use type_ahead_db::{execute, Pool, Query};
 
 // TODO move the handlers out
 async fn hello() -> impl Responder {
@@ -22,7 +20,7 @@ async fn execute_query(pool: &Pool, query: Query) -> Result<HttpResponse, actix_
     let pool = pool.clone();
     let conn = web::block(move || pool.get()).await?.unwrap(); // TODO: flat_map or similiar
 
-    let result = web::block(move || db::execute(conn, query))
+    let result = web::block(move || execute(conn, query))
         .await?
         .map_err(error::ErrorInternalServerError)?;
 
@@ -33,8 +31,7 @@ async fn find_cities_starting_with(
     pool: web::Data<Pool>,
     query_params: web::Query<TypeAheadParams>,
 ) -> Result<HttpResponse, actix_web::Error> {
-    let query =
-        db::Query::CityNamesStartingWith(query_params.prefix.to_string(), query_params.limit);
+    let query = Query::CityNamesStartingWith(query_params.prefix.to_string(), query_params.limit);
 
     execute_query(&pool, query).await
 }
@@ -44,7 +41,7 @@ async fn find_countries_starting_with(
     query_params: web::Query<TypeAheadParams>,
 ) -> Result<HttpResponse, actix_web::Error> {
     let query =
-        db::Query::CountryNamesStartingWith(query_params.prefix.to_string(), query_params.limit);
+        Query::CountryNamesStartingWith(query_params.prefix.to_string(), query_params.limit);
 
     execute_query(&pool, query).await
 }
@@ -53,8 +50,7 @@ async fn find_states_starting_with(
     pool: web::Data<Pool>,
     query_params: web::Query<TypeAheadParams>,
 ) -> Result<HttpResponse, actix_web::Error> {
-    let query =
-        db::Query::StateNamesStartingWith(query_params.prefix.to_string(), query_params.limit);
+    let query = Query::StateNamesStartingWith(query_params.prefix.to_string(), query_params.limit);
 
     execute_query(&pool, query).await
 }
